Split StreetEdgeTest into three
